requirements.txt
pytesseract
pillow
numpy
scipy
opencv-python
pytest
pytest-cov
unittest-mock